+++ pkg/server/api/release/routes.go
@@ -28,6 +28,6 @@ func Routers() []types.IRouter {
 		&public.Refresh{},
 		&public.CasdoorConfig{},
 		//&public.Statistics{},
-		&public.OpenAPI{},
+		&public.InternalOpenAPI{},
 	}
 }